Move permission create tests to a separate file
It's particularly painful to have too many Selenium tests in one
file, since each one takes an unavoidable amount of overhead to
set up the browser.  Separate the permission create tests from the
tests of list pagination and sorting, since they use different
pages and different setup data.
