+++ itests/fe/permissions_test.py
@@ -4,11 +4,8 @@ from typing import TYPE_CHECKING
 
 from pytz import UTC
 
-from grouper.constants import PERMISSION_CREATE
 from grouper.entities.permission import Permission
 from grouper.fe.settings import FrontendSettings
-from itests.pages.permission_create import PermissionCreatePage
-from itests.pages.permission_view import PermissionViewPage
 from itests.pages.permissions import PermissionsPage
 from itests.setup import frontend_server
 from tests.path_util import src_path
@@ -140,43 +137,3 @@ def test_list_pagination(tmpdir, setup, browser):
         seen_permissions = [(r.name, r.description, r.created_on) for r in page.permission_rows]
         assert seen_permissions == sorted(expected_permissions)[2:]
         assert page.limit_label == "Limit: 10"
-
-
-def test_list_create_button(tmpdir, setup, browser):
-    # type: (LocalPath, SetupTest, Chrome) -> None
-    with setup.transaction():
-        setup.create_user("gary@a.co")
-
-    with frontend_server(tmpdir, "gary@a.co") as frontend_url:
-        browser.get(url(frontend_url, "/permissions"))
-        page = PermissionsPage(browser)
-        assert not page.has_create_permission_button
-
-        with setup.transaction():
-            setup.grant_permission_to_group(PERMISSION_CREATE, "*", "admins")
-            setup.add_user_to_group("gary@a.co", "admins")
-        browser.get(url(frontend_url, "/permissions?refresh=yes"))
-        assert page.has_create_permission_button
-
-
-def test_create_permission(tmpdir, setup, browser):
-    # type: (LocalPath, SetupTest, Chrome) -> None
-    with setup.transaction():
-        setup.add_user_to_group("gary@a.co", "some-group")
-        setup.grant_permission_to_group(PERMISSION_CREATE, "foo.*", "some-group")
-        setup.grant_permission_to_group(PERMISSION_CREATE, "bar.baz", "some-group")
-
-    with frontend_server(tmpdir, "gary@a.co") as frontend_url:
-        browser.get(url(frontend_url, "/permissions"))
-        page = PermissionsPage(browser)
-        page.click_create_permission_button()
-
-        create_page = PermissionCreatePage(browser)
-        assert create_page.allowed_patterns == ["bar.baz", "foo.*"]
-        create_page.set_name("foo.bar")
-        create_page.set_description("testing")
-        create_page.form.submit()
-
-        view_page = PermissionViewPage(browser)
-        assert view_page.subheading == "foo.bar"
-        assert view_page.description == "testing"
